Admin
Working on admin panel

+++ admin/libraries/controllers/Comment.php
@@ -1,13 +1,19 @@
 <?php
 namespace Controllers;
 
-class Comment extends Controller
+abstract class Comment extends Controller
 {
     protected $modelName = "Comment";
+
     public function insert()
     {
         $articleModel = new \Models\Article();
 
+        /**
+         * 1. On vérifie que les données ont bien été envoyées en POST
+         * D'abord, on récupère les informations à partir du POST
+         * Ensuite, on vérifie qu'elles ne sont pas nulles
+         */
         // On commence par l'author
         $author = filter_input(INPUT_POST, 'author', FILTER_SANITIZE_SPECIAL_CHARS);
 
@@ -39,22 +45,32 @@ class Comment extends Controller
 
     public function delete()
     {
-        /** * Récupération du paramètre "id" en GET  */
+        /**
+         * 1. Récupération du paramètre "id" en GET
+         */
         $id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);
         if (!$id) {
             die("Ho ! Fallait préciser le paramètre id en GET !");
         }
-        /**  * Suppression réelle du commentaire
-         * On récupère l'identifiant de l'article avant de supprimer le commentaire
+
+        /**
+         * 3. Vérification de l'existence du commentaire
          */
         $commentaire = $this->model->find($id);
-        $article_id = $commentaire['article_id'];
+        if (!$commentaire) {
+            die("Aucun commentaire n'a l'identifiant $id !");
+        }
+
+        /**
+         * 4. Suppression réelle du commentaire
+         * On récupère l'identifiant de l'article avant de supprimer le commentaire
+         */
         $this->model->delete($id);
 
         /**
          * 5. Redirection vers l'article en question
          */
-        
+        $article_id = $commentaire['article_id'];
         \Http::redirect('index.php?controller=article&task=show&id=' . $article_id);
     }
 }
